FlaskUI/requirements.txt
Flask
requests
numpy
google-auth
gunicorn
